+++ src/modules/feature-modules/admin/pages/organisations/organisation-edit.component.spec.ts
@@ -13,8 +13,9 @@ import { AdminModule } from '@modules/feature-modules/admin/admin.module';
 
 import { PageOrganisationEditComponent } from './organisation-edit.component';
 
-import { OrganisationsService } from '@modules/feature-modules/admin/services/organisations.service';
+import { OrganisationsService as AdminOrganisationsService } from '@modules/feature-modules/admin/services/organisations.service';
 import { ServiceUsersService } from '@modules/feature-modules/admin/services/service-users.service';
+import { OrganisationsService } from '@modules/shared/services/organisations.service';
 
 
 describe('FeatureModules/Admin/Pages/Organisations/PageOrganisationEditComponent', () => {
@@ -27,6 +28,7 @@ describe('FeatureModules/Admin/Pages/Organisations/PageOrganisationEditComponent
   let authenticationStore: AuthenticationStore;
   let serviceUserService: ServiceUsersService;
   let organisationsService: OrganisationsService;
+  let adminOrganisationsService: AdminOrganisationsService;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
@@ -158,7 +160,7 @@ describe('FeatureModules/Admin/Pages/Organisations/PageOrganisationEditComponent
   it('should run onSubmitWizard() with API success when updating organisation', () => {
     activatedRoute.snapshot.data = { module: 'Organisation' };
     authenticationStore.initializeAuthentication$ = () => of(true);
-    organisationsService.updateOrganisation = () => of({ id: 'Org01', status: 'OK' });
+    adminOrganisationsService.updateOrganisation = () => of({ id: 'Org01', status: 'OK' });
 
     fixture = TestBed.createComponent(PageOrganisationEditComponent);
     component = fixture.componentInstance;
@@ -172,7 +174,7 @@ describe('FeatureModules/Admin/Pages/Organisations/PageOrganisationEditComponent
     activatedRoute.snapshot.params = { organisationId: 'Org01', organisationUnitId: 'Unit01' };
     activatedRoute.snapshot.data = { module: 'Unit' };
     authenticationStore.initializeAuthentication$ = () => of(true);
-    organisationsService.updateUnit = () => of({ id: 'Unit01', status: 'OK' });
+    adminOrganisationsService.updateUnit = () => of({ id: 'Unit01', status: 'OK' });
 
     fixture = TestBed.createComponent(PageOrganisationEditComponent);
     component = fixture.componentInstance;
@@ -188,7 +190,7 @@ describe('FeatureModules/Admin/Pages/Organisations/PageOrganisationEditComponent
     activatedRoute.snapshot.data = { module: 'Organisation' };
     authenticationStore.initializeAuthentication$ = () => of(true);
 
-    organisationsService.updateOrganisation = () => throwError({ id: '123456ABCDFG' });
+    adminOrganisationsService.updateOrganisation = () => throwError({ id: '123456ABCDFG' });
 
     fixture = TestBed.createComponent(PageOrganisationEditComponent);
     component = fixture.componentInstance;
@@ -204,7 +206,7 @@ describe('FeatureModules/Admin/Pages/Organisations/PageOrganisationEditComponent
     activatedRoute.snapshot.data = { module: 'Unit' };
     authenticationStore.initializeAuthentication$ = () => of(true);
 
-    organisationsService.updateUnit = () => throwError({ id: '123456ABCDFG' });
+    adminOrganisationsService.updateUnit = () => throwError({ id: '123456ABCDFG' });
 
     fixture = TestBed.createComponent(PageOrganisationEditComponent);
     component = fixture.componentInstance;
@@ -217,7 +219,7 @@ describe('FeatureModules/Admin/Pages/Organisations/PageOrganisationEditComponent
   it('should run onSubmitWizard() with API error when updating organisation', () => {
     activatedRoute.snapshot.params = { organisationId: 'Org01' };
     activatedRoute.snapshot.data = { module: 'Organisation' };
-    organisationsService.updateOrganisation = () => throwError('error');
+    adminOrganisationsService.updateOrganisation = () => throwError('error');
 
     fixture = TestBed.createComponent(PageOrganisationEditComponent);
     component = fixture.componentInstance;
@@ -230,7 +232,7 @@ describe('FeatureModules/Admin/Pages/Organisations/PageOrganisationEditComponent
   it('should run onSubmitWizard() with API error when updating unit', () => {
     activatedRoute.snapshot.params = { organisationId: 'Org01', organisationUnitId: 'Unit01' };
     activatedRoute.snapshot.data = { module: 'Unit' };
-    organisationsService.updateUnit = () => throwError('error');
+    adminOrganisationsService.updateUnit = () => throwError('error');
 
     fixture = TestBed.createComponent(PageOrganisationEditComponent);
     component = fixture.componentInstance;
