+++ src/modules/feature-modules/admin/pages/service-users/service-user-info.component.spec.ts
@@ -14,7 +14,7 @@ import { AdminModule } from '@modules/feature-modules/admin/admin.module';
 import { PageServiceUserInfoComponent } from './service-user-info.component';
 
 import { getUserFullInfoDTO, ServiceUsersService } from '@modules/feature-modules/admin/services/service-users.service';
-import { OrganisationsService } from '@modules/feature-modules/admin/services/organisations.service';
+import { OrganisationsService } from '@modules/shared/services/organisations.service';
 import { AlertType } from '@app/base/types';
