+++ src/modules/feature-modules/admin/services/organisations.service.ts
@@ -4,33 +4,10 @@ import { catchError, map, take } from 'rxjs/operators';
 
 import { CoreService } from '@app/base';
 import { UrlModel } from '@app/base/models';
-import { APIQueryParamsType, DateISOType, MappedObjectType } from '@app/base/types';
+import { MappedObjectType } from '@app/base/types';
 
 import { AccessorOrganisationRoleEnum } from '@modules/stores/authentication/authentication.enums';
-import { InnovationSupportStatusEnum } from '@modules/stores/innovation';
 
-
-export type GetOrganisationsListDTO = {
-  id: string;
-  name: string;
-  acronym: string;
-  isActive: boolean;
-  organisationUnits: { id: string; name: string; acronym: string; isActive: boolean }[];
-};
-
-export type GetOrganisationInfoDTO = {
-  id: string;
-  name: string;
-  acronym: string;
-  isActive: boolean;
-  organisationUnits: {
-    id: string;
-    name: string;
-    acronym: string;
-    isActive: boolean;
-    userCount: number;
-  }[];
-};
 
 export type updateOrganisationDTO = {
   id: string;
@@ -38,42 +15,6 @@ export type updateOrganisationDTO = {
   error?: string;
 };
 
-export type GetOrganisationUnitInfoDTO = {
-  id: string;
-  name: string;
-  acronym: string;
-  isActive: boolean;
-  userCount: number;
-};
-
-export type GetOrganisationUnitUsersInDTO = {
-  count: number;
-  data: {
-    id: string;
-    name: string;
-    email: string;
-    organisationRole: AccessorOrganisationRoleEnum;
-    isActive: boolean;
-    lockedAt: null | DateISOType;
-  }[];
-};
-export type GetOrganisationUnitUsersOutDTO = {
-  count: number;
-  data: (GetOrganisationUnitUsersInDTO['data'][0] & { organisationRoleDescription: string })[];
-};
-
-export type GetOrganisationUnitInnovationsListDTO = {
-  count: number;
-  innovationsByStatus: {
-    status: InnovationSupportStatusEnum,
-    count: number
-  }[];
-  innovationsList: {
-    id: string,
-    name: string,
-    status: InnovationSupportStatusEnum
-  }[];
-};
 
 export type organisationUsersInDTO = {
   id: string;
@@ -100,33 +41,6 @@ export class OrganisationsService extends CoreService {
 
   constructor() { super(); }
 
-  getOrganisationsList(filters: { withInactive: boolean }): Observable<GetOrganisationsListDTO[]> {
-
-    const url = new UrlModel(this.API_USERS_URL).addPath('v1/organisations').setQueryParams({
-      withInactive: filters.withInactive,
-      fields: ['organisationUnits'] // always get the organisation units to keep previous behaviour
-    });
-    return this.http.get<GetOrganisationsListDTO[]>(url.buildUrl()).pipe(
-      take(1),
-      map(response => response)
-    );
-
-  }
-
-  getOrganisationInfo(organisationId: string): Observable<GetOrganisationInfoDTO> {
-
-    const url = new UrlModel(this.API_USERS_URL).addPath('v1/organisations/:organisationId').setPathParams({ organisationId });
-    return this.http.get<GetOrganisationInfoDTO>(url.buildUrl()).pipe(take(1),
-      map(response => ({
-        id: response.id, name: response.name, acronym: response.acronym, isActive: response.isActive,
-        organisationUnits: response.organisationUnits.map(item => ({
-          id: item.id, name: item.name, acronym: item.acronym, isActive: item.isActive, userCount: item.userCount
-        }))
-      }))
-    );
-
-  }
-
   updateOrganisation(body: MappedObjectType, securityConfirmation: { id: string, code: string }, organisationId: string): Observable<updateOrganisationDTO> {
 
     const qp = (securityConfirmation.id && securityConfirmation.code) ? securityConfirmation : {};
@@ -140,49 +54,6 @@ export class OrganisationsService extends CoreService {
 
   }
 
-
-  getOrganisationUnitInfo(organisationId: string, organisationUnitId: string): Observable<GetOrganisationUnitInfoDTO> {
-
-    const url = new UrlModel(this.API_USERS_URL).addPath('v1/organisations/:organisationId/units/:organisationUnitId').setPathParams({ organisationId, organisationUnitId });
-    return this.http.get<GetOrganisationUnitInfoDTO>(url.buildUrl()).pipe(take(1),
-      map(response => response)
-    );
-
-  }
-
-  getOrganisationUnitUsers(organisationId: string, organisationUnitId: string, queryParams: APIQueryParamsType<{ onlyActive: boolean }>): Observable<GetOrganisationUnitUsersOutDTO> {
-
-    const { filters, ...qParams } = queryParams;
-    const qp = {
-      ...qParams,
-      onlyActive: filters.onlyActive ? 'true' : 'false'
-    };
-
-    const url = new UrlModel(this.API_URL).addPath('user-admin/organisations/:organisationId/units/:organisationUnitId/users').setPathParams({ organisationId, organisationUnitId }).setQueryParams(qp);
-    return this.http.get<GetOrganisationUnitUsersInDTO>(url.buildUrl()).pipe(
-      take(1),
-      map(response => ({
-        count: response.count,
-        data: response.data.map(user => ({ ...user, organisationRoleDescription: this.stores.authentication.getRoleDescription(user.organisationRole) }))
-      }))
-    );
-
-  }
-
-  getOrganisationUnitInnovationsList(organisationId: string, organisationUnitId: string, queryParams: APIQueryParamsType<{ onlyOpen: boolean }>): Observable<GetOrganisationUnitInnovationsListDTO> {
-
-    const { filters, ...qParams } = queryParams;
-    const qp = {
-      ...qParams,
-      onlyOpen: filters.onlyOpen ? 'true' : 'false'
-    };
-
-    const url = new UrlModel(this.API_URL).addPath('user-admin/organisations/:organisationId/units/:organisationUnitId/innovations').setPathParams({ organisationId, organisationUnitId }).setQueryParams(qp);
-    return this.http.get<GetOrganisationUnitInnovationsListDTO>(url.buildUrl()).pipe(take(1),
-      map(response => response)
-    );
-
-  }
 
   updateUnit(body: MappedObjectType, securityConfirmation: { id: string, code: string }, organisationUnitId: string, organisationId: string): Observable<updateOrganisationDTO> {
