+++ src/modules/shared/services/organisations.service.ts
@@ -6,6 +6,8 @@ import { CoreService } from '@app/base';
 import { UserRoleEnum } from '@app/base/enums';
 import { UrlModel } from '@app/base/models';
 import { UserSearchDTO } from '../dtos/users.dto';
+import { InnovationSupportStatusEnum } from '@modules/stores/innovation';
+import { APIQueryParamsType } from '@app/base/types';
 
 
 export type getAccessorsOrganisationsDTO = {
@@ -17,7 +19,7 @@ export type OrganisationsListDTO = {
   id: string,
   name: string,
   acronym: string,
-  organisationUnits: { id: string, name: string, acronym: string }[];
+  organisationUnits: { id: string, name: string, acronym: string, isActive: boolean }[];
 };
 
 export type GetOrganisationInfoDTO = {
@@ -52,6 +54,19 @@ export type GetOrganisationUnitUsersDTO = {
   isActive: boolean,
   lockedAt: string | undefined
 }[];
+
+export type GetOrganisationUnitInnovationsListDTO = {
+  count: number;
+  innovationsByStatus: {
+    status: InnovationSupportStatusEnum,
+    count: number
+  }[];
+  innovationsList: {
+    id: string,
+    name: string,
+    status: InnovationSupportStatusEnum
+  }[];
+};
 
 @Injectable()
 export class OrganisationsService extends CoreService {
@@ -128,6 +143,22 @@ export class OrganisationsService extends CoreService {
     return this.http.get<GetOrganisationUnitInfoDTO>(url.buildUrl()).pipe(take(1),
       map(response => response)
     );
+  }
+
+  
+  getOrganisationUnitInnovationsList(organisationId: string, organisationUnitId: string, queryParams: APIQueryParamsType<{ onlyOpen: boolean }>): Observable<GetOrganisationUnitInnovationsListDTO> {
+
+    const { filters, ...qParams } = queryParams;
+    const qp = {
+      ...qParams,
+      onlyOpen: filters.onlyOpen ? 'true' : 'false'
+    };
+
+    const url = new UrlModel(this.API_URL).addPath('user-admin/organisations/:organisationId/units/:organisationUnitId/innovations').setPathParams({ organisationId, organisationUnitId }).setQueryParams(qp);
+    return this.http.get<GetOrganisationUnitInnovationsListDTO>(url.buildUrl()).pipe(take(1),
+      map(response => response)
+    );
+
   }
 
 }
